generalize methods
